+++ bpf/insn/samples/loop.c
@@ -1,6 +1,8 @@
 /**
- * long bpf_loop(__u32 nr_loops, void *callback_fn, void *callback_ctx, __u64 flags);
+ * long bpf_loop(__u32 nr_loops, void *callback_fn, void *callback_ctx,
+ *               __u64 flags);
  * BPF_FUNC_loop = 181;
+ * typedef int (*callback_fn)(__u32 index, void *callback_ctx);
  *
  * Refs:
  * - https://docs.ebpf.io/linux/helper-function/bpf_loop/
@@ -8,11 +10,14 @@
 #include "bpf_insn_samples.h"
 
 static struct bpf_insn insns[] = {
+	BPF_MOV64_IMM(BPF_REG_1, 10),
+	BPF_MOV64_IMM(BPF_REG_2, 0), /* TODO: callback_fn */
+	BPF_MOV64_IMM(BPF_REG_3, 0),
+	BPF_MOV64_IMM(BPF_REG_4, 0),
 	BPF_RAW_INSN(BPF_JMP | BPF_CALL, 0, 0, 0, BPF_FUNC_loop),
 	BPF_MOV64_IMM(BPF_REG_0, 0),
 	BPF_EXIT_INSN(),
 };
-
 
 BPF_INSN_SAMPLE_FUNC_PROTO(loop)
 {